app/SmartVisitorSystem/requirements.txt
opencv-python
pillow
openpyxl
tkinter
numpy
